refactor(server): peel offload routes into routers/offload.py (router-split, round-5 #51 / R5-25 PR18)
Seventh router peeled (merge-as-green off main after bins/PR17). The /api/offload
group — layout preview + streaming copy+verify run (never deletes source) — plus
the legacy /dit redirect move into routers/offload.py, carrying the 2 body models
and the per-source single-flight primitives (_offload_lock / _offload_active /
_acquire_offload_slot / _release_offload_slot, R5-17 #19; offload-only, keyed on
resolved source path). BASE_DIR (config) replaces server.ROOT for offload.py.
Verbatim subprocess/streaming/GeneratorExit/resume-state logic. Imports auth +
config + webguard (dst denylist) + offload — no server import.

test_r5_17_offload_lifecycle referenced server._acquire_offload_slot to simulate
an in-flight run; repointed to routers.offload (the slot's new home) so the
in-flight simulation still blocks the route (409). New boundary test pins
leaf/ownership/mount + slot guard semantics. 864 passed.
